# catalog/internal/server/openapi/api_model_catalog_service.go
/*
 * Model Catalog REST API
 *
 * REST API for Model Registry to create and manage ML model metadata
 *
 * API version: v1alpha1
 */

// Generated by: OpenAPI Generator (https://openapi-generator.tech). DO NOT EDIT.

package openapi

import (
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"

	model "github.com/kubeflow/model-registry/catalog/pkg/openapi"
)

// ModelCatalogServiceAPIController binds http requests to an api service and writes the service results to the http response
type ModelCatalogServiceAPIController struct {
	service      ModelCatalogServiceAPIServicer
	errorHandler ErrorHandler
}

// ModelCatalogServiceAPIOption for how the controller is set up.
type ModelCatalogServiceAPIOption func(*ModelCatalogServiceAPIController)

// WithModelCatalogServiceAPIErrorHandler inject ErrorHandler into controller
func WithModelCatalogServiceAPIErrorHandler(h ErrorHandler) ModelCatalogServiceAPIOption {
	return func(c *ModelCatalogServiceAPIController) {
		c.errorHandler = h
	}
}

// NewModelCatalogServiceAPIController creates a default api controller
func NewModelCatalogServiceAPIController(s ModelCatalogServiceAPIServicer, opts ...ModelCatalogServiceAPIOption) Router {
	controller := &ModelCatalogServiceAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the ModelCatalogServiceAPIController
func (c *ModelCatalogServiceAPIController) Routes() Routes {
	return Routes{
		"FindModels": Route{
			strings.ToUpper("Get"),
			"/api/model_catalog/v1alpha1/models",
			c.FindModels,
		},
		"FindSources": Route{
			strings.ToUpper("Get"),
			"/api/model_catalog/v1alpha1/sources",
			c.FindSources,
		},
		"GetModel": Route{
			strings.ToUpper("Get"),
			"/api/model_catalog/v1alpha1/sources/{source_id}/models/*",
			c.GetModel,
		},
		"GetAllModelArtifacts": Route{
			strings.ToUpper("Get"),
			"/api/model_catalog/v1alpha1/sources/{source_id}/models/{model_name}/artifacts",
			c.GetAllModelArtifacts,
		},
	}
}

// FindModels - Search catalog models across sources.
func (c *ModelCatalogServiceAPIController) FindModels(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	ownerParam := query.Get("owner")
	sourceParam := query.Get("source")
	qParam := query.Get("q")
	pageSizeParam := query.Get("pageSize")
	orderByParam := query.Get("orderBy")
	sortOrderParam := query.Get("sortOrder")
	nextPageTokenParam := query.Get("nextPageToken")
	userIdParam := query.Get("userId")
	result, err := c.service.FindModels(r.Context(), ownerParam, sourceParam, qParam, pageSizeParam, model.OrderByField(orderByParam), model.SortOrder(sortOrderParam), nextPageTokenParam, userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// FindSources - List All CatalogSources
func (c *ModelCatalogServiceAPIController) FindSources(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	ownerParam := query.Get("owner")
	nameParam := query.Get("name")
	pageSizeParam := query.Get("pageSize")
	orderByParam := query.Get("orderBy")
	sortOrderParam := query.Get("sortOrder")
	nextPageTokenParam := query.Get("nextPageToken")
	userIdParam := query.Get("userId")
	result, err := c.service.FindSources(r.Context(), ownerParam, nameParam, pageSizeParam, model.OrderByField(orderByParam), model.SortOrder(sortOrderParam), nextPageTokenParam, userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetModel - Get a `CatalogModel`.
func (c *ModelCatalogServiceAPIController) GetModel(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	sourceIdParam := chi.URLParam(r, "source_id")
	modelNameParam := chi.URLParam(r, "*")
	ownerParam := query.Get("owner")
	userIdParam := query.Get("userId")
	result, err := c.service.GetModel(r.Context(), sourceIdParam, modelNameParam, ownerParam, userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetAllModelArtifacts - List CatalogModelArtifacts.
func (c *ModelCatalogServiceAPIController) GetAllModelArtifacts(w http.ResponseWriter, r *http.Request) {
	query := r.URL.Query()
	sourceIdParam := chi.URLParam(r, "source_id")
	modelNameParam := chi.URLParam(r, "model_name")
	ownerParam := query.Get("owner")
	userIdParam := query.Get("userId")
	result, err := c.service.GetAllModelArtifacts(r.Context(), sourceIdParam, modelNameParam, ownerParam, userIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)
}